requirements.txt
# Foreman v1.2 FIXED - Dipendenze EndeavourOS Verificate
# Compatibilità testata: Python 3.13, EndeavourOS 2025

# ADK Core (versione stabile con fix MCP)
google-adk>=1.1.0,<1.3.0

# Environment management
python-dotenv>=1.0.0

# MCP Support (versione compatibile ADK)
mcp>=1.10.0,<2.0.0

# Additional dependencies che potrebbero servire
pydantic>=2.0.0
httpx>=0.25.0
anyio>=4.0.0

# Development e debugging (opzionali)
# ipython>=8.0.0
# pytest>=7.0.0
